pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "detective-benno"
version = "0.1.0"
description = "An intelligent code review detective powered by LLM - solving code mysteries, one PR at a time"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Bima Kharisma Wicaksana", email = "bimakw@example.com" }
]
keywords = [
    "code-review",
    "ai",
    "llm",
    "openai",
    "github",
    "automation",
    "devtools",
    "detective",
    "security",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "pygments>=2.15.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
github = [
    "PyGithub>=2.0.0",
]

[project.scripts]
benno = "detective_benno.cli:main"

[project.urls]
Homepage = "https://github.com/bimakw/detective-benno"
Documentation = "https://github.com/bimakw/detective-benno#readme"
Repository = "https://github.com/bimakw/detective-benno.git"
Issues = "https://github.com/bimakw/detective-benno/issues"
Changelog = "https://github.com/bimakw/detective-benno/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/detective_benno"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["detective_benno"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/detective_benno"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
